@@ -51,31 +58,33 @@
   },
   "homepage": "https://github.com/martinroob/tiny-translator#readme",
   "dependencies": {
-    "@angular/animations": "^5.0.3",
-    "@angular/cdk": "^5.0.0-rc.2",
-    "@angular/common": "^5.0.3",
-    "@angular/compiler": "^5.0.3",
-    "@angular/core": "^5.0.3",
-    "@angular/flex-layout": "^5.0.0-beta.15",
-    "@angular/forms": "^5.0.3",
-    "@angular/http": "^5.0.3",
-    "@angular/material": "^5.0.0-rc.2",
-    "@angular/platform-browser": "^5.0.3",
-    "@angular/platform-browser-dynamic": "^5.0.3",
-    "@angular/router": "^5.0.3",
-    "@angular/service-worker": "^5.0.3",
+    "@angular/animations": "^7.2.15",
+    "@angular/cdk": "^7.3.7",
+    "@angular/common": "^7.2.15",
+    "@angular/compiler": "^7.2.15",
+    "@angular/core": "^7.2.15",
+    "@angular/flex-layout": "^7.0.0-beta.24",
+    "@angular/forms": "^7.2.15",
+    "@angular/http": "^7.2.15",
+    "@angular/material": "^7.3.7",
+    "@angular/platform-browser": "^7.2.15",
+    "@angular/platform-browser-dynamic": "^7.2.15",
+    "@angular/router": "^7.2.15",
+    "@angular/service-worker": "^7.2.15",
+    "angular-cli-ghpages": "^0.5.3",
     "core-js": "^2.4.1",
     "file-saver": "^1.3.3",
     "flag-icon-css": "^2.8.0",
     "hammerjs": "^2.0.8",
-    "ngx-i18nsupport-lib": "^1.6.1",
-    "rxjs": "^5.5.2",
-    "zone.js": "^0.8.14"
+    "ngx-i18nsupport-lib": "^1.7.0",
+    "rxjs": "^6.5.2",
+    "zone.js": "^0.8.27"
   },
   "devDependencies": {
-    "@angular/cli": "^1.5.4",
-    "@angular/compiler-cli": "^5.0.3",
-    "@angular/language-service": "^5.0.3",
+    "@angular-devkit/build-angular": "~0.8.0",
+    "@angular/cli": "^7.3.7",
+    "@angular/compiler-cli": "^7.2.15",
+    "@angular/language-service": "^7.2.15",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
@@ -90,11 +99,11 @@
     "karma-coverage-istanbul-reporter": "^1.2.1",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "ngx-i18nsupport": "^0.8.4",
+    "ngx-i18nsupport": "^0.17.1",
     "protractor": "~5.1.2",
     "replace": "^0.3.0",
-    "ts-node": "~3.2.0",
-    "tslint": "^5.8.0",
-    "typescript": "2.4"
+    "ts-node": "~8.0.0",
+    "tslint": "^5.18.0",
+    "typescript": "3.2.1"
   }
 }
