@@ -1,4 +1,4 @@
-import { Observable } from 'rxjs/Observable';
+import { throwError as observableThrowError, Observable, of } from 'rxjs';
 
 /**
  * Reasons, why you cannot use the API.
@@ -41,7 +41,7 @@ export class AutoTranslateServiceAPI {
    * @param target the language to translate to
    */
   public canAutoTranslate(source: string, target: string): Observable<boolean> {
-    return Observable.of(false);
+    return of(false);
   }
 
   /**
@@ -51,7 +51,7 @@ export class AutoTranslateServiceAPI {
    * @return {AutoTranslateDisabledReason} or null, if API is enabled.
    */
   public disabledReason(source: string, target: string): Observable<AutoTranslateDisabledReason> {
-    return Observable.of({
+    return of({
       reason: AutoTranslateDisabledReasonKey.NO_PROVIDER,
     });
   }
@@ -62,7 +62,7 @@ export class AutoTranslateServiceAPI {
    * @param target language for readable name. (default is en)
    */
   supportedLanguages(target?: string): Observable<Language[]> {
-    return Observable.of([]);
+    return of([]);
   }
 
   /**
@@ -74,7 +74,7 @@ export class AutoTranslateServiceAPI {
    * @return Observable with translated message or error
    */
   public translate(message: string, from: string, to: string): Observable<string> {
-    return Observable.throw('no translation service installed');
+    return observableThrowError('no translation service installed');
   }
 
   /**
@@ -85,6 +85,6 @@ export class AutoTranslateServiceAPI {
    * @return Observable with translated messages or error
    */
   public translateMultipleStrings(messages: string[], from: string, to: string): Observable<string[]> {
-    return Observable.throw('no translation service installed');
+    return observableThrowError('no translation service installed');
   }
 }
