@@ -3,13 +3,14 @@ import { TranslationUnit } from '../model/translation-unit';
 import { MatDialog, MatSnackBar } from '@angular/material';
 import { NormalizedMessage } from '../model/normalized-message';
 import { FormBuilder, FormGroup } from '@angular/forms';
-import { Observable } from 'rxjs/Observable';
+import { Observable, of } from 'rxjs';
 import { TranslateUnitWarningConfirmDialogComponent } from '../translate-unit-warning-confirm-dialog/translate-unit-warning-confirm-dialog.component';
 import { TranslationFileView } from '../model/translation-file-view';
 import { WorkflowType } from '../model/translation-project';
 import { STATE_FINAL, STATE_TRANSLATED } from 'ngx-i18nsupport-lib/dist';
 import { AutoTranslateServiceAPI } from '../model/auto-translate-service-api';
 import { isNullOrUndefined } from 'util';
+import { map } from 'rxjs/operators';
 
 export enum NavigationDirection {
   NEXT,
@@ -294,9 +295,9 @@ export class TranslateUnitComponent implements OnInit, OnChanges {
     const errors = this.errors();
     if (warnings.length === 0 && errors.length === 0) {
       // everything good, we don´t need a dialog then.
-      return Observable.of('accept');
+      return of('accept');
     } else if (!this.isTranslationChanged()) {
-      return Observable.of('accept');
+      return of('accept');
     } else {
       const dialogRef = this.dialog.open(TranslateUnitWarningConfirmDialogComponent, {
         data: { errors: errors, warnings: warnings },
@@ -391,10 +392,10 @@ export class TranslateUnitComponent implements OnInit, OnChanges {
 
   autoTranslateDisabled(): Observable<boolean> {
     if (!this.translationUnit) {
-      return Observable.of(true);
+      return of(true);
     }
     return this.autoTranslateService
       .canAutoTranslate(this.translationUnit.translationFile().sourceLanguage(), this.translationUnit.translationFile().targetLanguage())
-      .map(val => !val);
+      .pipe(map(val => !val));
   }
 }
