@@ -3,7 +3,8 @@ import { NormalizedMessage } from '../model/normalized-message';
 import { isNullOrUndefined } from 'util';
 import { ControlValueAccessor, FormBuilder, FormGroup, NG_VALUE_ACCESSOR } from '@angular/forms';
 import { IICUMessageCategory, IICUMessageTranslation } from 'ngx-i18nsupport-lib/dist';
-import { Subscription } from 'rxjs/Subscription';
+import { Subscription } from 'rxjs';
+import { debounceTime } from 'rxjs/operators';
 /**
  * A component used as an input field for normalized message.
  */
@@ -72,7 +73,7 @@ export class NormalizedMessageInputComponent implements OnInit, OnChanges, Contr
       displayedText: [{ value: this.textToDisplay(), disabled: this.disabled }],
       icuMessages: this.formBuilder.array(this.initIcuMessagesFormArray()),
     });
-    this.subscription = this.form.valueChanges.debounceTime(200).subscribe(formValue => {
+    this.subscription = this.form.valueChanges.pipe(debounceTime(200)).subscribe(formValue => {
       this.valueChanged(formValue);
     });
   }
